201906151759 修正TC数据类型未float32 增加任务启动延迟

+++ object/Object.go
@@ -32,7 +32,7 @@ type ZlCompany struct {
 type MdYyInfo struct {
 	FMdId    int
 	FYyr     time.Time
-	FTc      int
+	FTc      float32
 	FSr      float32
 	FOprTime time.Time
 }
